server/pyproject.toml
[tool.poetry]
name = "check-my-plants"
version = "0.1.0"
description = ""
authors = ["tobiwankenobii <tobiv1337@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
pydantic = {extras = ["email"], version = "^1.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
httpx = "^0.18.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.3"}
aiofiles = "^0.7.0"
Jinja2 = "^3.0.1"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b1"
isort = "^5.8.0"
flake8 = "^3.9.2"
pytest-asyncio = "^0.15.1"
pytest-cov = "^2.12.0"
asgi-lifespan = "^1.0.1"
nest-asyncio = "^1.5.1"
asynctest = "^0.13.0"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
